fix: terminal cwd tests
